refactor: move pkg types to ceresdb
