+++ ceresdb/client_impl.go
@@ -5,17 +5,14 @@ package ceresdb
 import (
 	"context"
 	"fmt"
-
-	"github.com/CeresDB/ceresdb-client-go/types"
-	"github.com/CeresDB/ceresdb-client-go/utils"
 )
 
 type clientImpl struct {
 	rpcClient   *rpcClient
-	routeClient RouteClient
+	routeClient routeClient
 }
 
-func newClient(endpoint string, routeMode types.RouteMode, opts options) (Client, error) {
+func newClient(endpoint string, routeMode RouteMode, opts options) (Client, error) {
 	rpcClient := newRPCClient(opts)
 	routeClient, err := newRouteClient(endpoint, routeMode, rpcClient, opts)
 	if err != nil {
@@ -27,69 +24,69 @@ func newClient(endpoint string, routeMode types.RouteMode, opts options) (Client
 	}, nil
 }
 
-func (c *clientImpl) SQLQuery(ctx context.Context, req types.SQLQueryRequest) (types.SQLQueryResponse, error) {
+func (c *clientImpl) SQLQuery(ctx context.Context, req SQLQueryRequest) (SQLQueryResponse, error) {
 	if err := c.withDefaultRequestContext(&req.ReqCtx); err != nil {
-		return types.SQLQueryResponse{}, err
+		return SQLQueryResponse{}, err
 	}
 
 	if len(req.Tables) == 0 {
-		return types.SQLQueryResponse{}, types.ErrNullRequestTables
+		return SQLQueryResponse{}, ErrNullRequestTables
 	}
 
 	routes, err := c.routeClient.RouteFor(req.ReqCtx, req.Tables)
 	if err != nil {
-		return types.SQLQueryResponse{}, fmt.Errorf("Route tables failed, tables:%v, err:%v", req.Tables, err)
+		return SQLQueryResponse{}, fmt.Errorf("route tables failed, tables:%v, err:%v", req.Tables, err)
 	}
 	for _, route := range routes {
 		queryResponse, err := c.rpcClient.SQLQuery(ctx, route.Endpoint, req)
-		if ceresdbErr, ok := err.(*types.CeresdbError); ok && ceresdbErr.ShouldClearRoute() {
+		if ceresdbErr, ok := err.(*CeresdbError); ok && ceresdbErr.ShouldClearRoute() {
 			c.routeClient.ClearRouteFor(req.Tables)
 		}
 		return queryResponse, err
 	}
-	return types.SQLQueryResponse{}, types.ErrEmptyRoute
+	return SQLQueryResponse{}, ErrEmptyRoute
 }
 
-func (c *clientImpl) Write(ctx context.Context, req types.WriteRequest) (types.WriteResponse, error) {
+func (c *clientImpl) Write(ctx context.Context, req WriteRequest) (WriteResponse, error) {
 	if err := c.withDefaultRequestContext(&req.ReqCtx); err != nil {
-		return types.WriteResponse{}, err
+		return WriteResponse{}, err
 	}
 
 	if len(req.Points) == 0 {
-		return types.WriteResponse{}, types.ErrNullRows
+		return WriteResponse{}, ErrNullRows
 	}
 
-	tables := utils.GetTablesFromPoints(req.Points)
+	tables := getTablesFromPoints(req.Points)
 
 	routes, err := c.routeClient.RouteFor(req.ReqCtx, tables)
 	if err != nil {
-		return types.WriteResponse{}, err
+		return WriteResponse{}, err
 	}
 
-	pointsByRoute, err := utils.SplitPointsByRoute(req.Points, routes)
+	pointsByRoute, err := splitPointsByRoute(req.Points, routes)
 	if err != nil {
-		return types.WriteResponse{}, err
+		return WriteResponse{}, err
 	}
 
 	// TODO
 	// Convert to parallel write
-	ret := types.WriteResponse{}
+	ret := WriteResponse{}
 	for endpoint, points := range pointsByRoute {
 		response, err := c.rpcClient.Write(ctx, endpoint, req.ReqCtx, points)
 		if err != nil {
-			if ceresdbErr, ok := err.(*types.CeresdbError); ok && ceresdbErr.ShouldClearRoute() {
-				c.routeClient.ClearRouteFor(utils.GetTablesFromPoints(points))
+			if ceresdbErr, ok := err.(*CeresdbError); ok && ceresdbErr.ShouldClearRoute() {
+				c.routeClient.ClearRouteFor(getTablesFromPoints(points))
 			}
 
-			ret = utils.CombineWriteResponse(ret, types.WriteResponse{Failed: uint32(len(points))})
+			ret = combineWriteResponse(ret, WriteResponse{Failed: uint32(len(points))})
 			continue
 		}
-		ret = utils.CombineWriteResponse(ret, response)
+		ret = combineWriteResponse(ret, response)
 	}
 	return ret, nil
 }
 
-func (c *clientImpl) withDefaultRequestContext(reqCtx *types.RequestContext) error {
+func (c *clientImpl) withDefaultRequestContext(reqCtx *RequestContext) error {
 	// use default
 	if reqCtx.Database == "" {
 		reqCtx.Database = c.rpcClient.opts.Database
@@ -97,7 +94,7 @@ func (c *clientImpl) withDefaultRequestContext(reqCtx *types.RequestContext) err
 
 	// check Request Context
 	if reqCtx.Database == "" {
-		return types.ErrNoDatabaseSelected
+		return ErrNoDatabaseSelected
 	}
 	return nil
 }
