--- types/common.go
@@ -1,7 +1,0 @@
-// Copyright 2022 CeresDB Project Authors. Licensed under Apache-2.0.
-
-package types
-
-type RequestContext struct {
-	Database string
-}
